dev to main

+++ fitDash/amqp/consumer.py
@@ -10,9 +10,10 @@ RABBITMQ_DEFAULT_HOST = os.getenv("RABBITMQ_DEFAULT_HOST")
 RABBITMQ_DEFAULT_USER = os.getenv("RABBITMQ_DEFAULT_USER")
 RABBITMQ_DEFAULT_PASS = os.getenv("RABBITMQ_DEFAULT_PASS")
 RABBITMQ_DEFAULT_VHOST = os.getenv("RABBITMQ_DEFAULT_VHOST")
-RABBITMQ_QUEUE_NAME = os.getenv("RABBITMQ_QUEUE_NAME")
+RABBITMQ_DEFAULT_PORT = os.getenv("RABBITMQ_DEFAULT_PORT")
 RABBITMQ_EXCHANGE_NAME = os.getenv("RABBITMQ_EXCHANGE_NAME")
 RABBITMQ_EXCHANGE_TYPE = os.getenv("RABBITMQ_EXCHANGE_TYPE")
+RABBITMQ_QUEUE_NAME=os.getenv("RABBITMQ_QUEUE_NAME")
 
 os.environ.setdefault("DJANGOSETTINGS_MODULE", "fitDash.settings")
 django.setup()
@@ -22,6 +23,7 @@ def __get_connection_and_channel():
     credentials = pika.PlainCredentials(RABBITMQ_DEFAULT_USER, RABBITMQ_DEFAULT_PASS)
     parameters = pika.ConnectionParameters(
         host=RABBITMQ_DEFAULT_HOST,
+        port=int(RABBITMQ_DEFAULT_PORT),
         virtual_host=RABBITMQ_DEFAULT_VHOST,
         credentials=credentials,
     )
@@ -49,7 +51,8 @@ def start_delete_user_objects():
                 try:
                     data = json.loads(body)
                     user_id = data.get("user_id")
-                    if user_id is not None:
+                    event=data.get("event")
+                    if user_id is not None and event=="delete":
                         Habit.objects.filter(user_id=user_id).delete()
                         ToDo.objects.filter(user_id=user_id).delete()
                 except Exception as e:
